feat: deleteBoard 라우트 핸들러, 서비스 로직 수정
deleteBoard 라우트 핸들러
  - 파라미터에 request 객체 추가

deleteBoard 서비스 로직
  - 파라미터에 userId 추가
  - delete 조건에 userId 추가

board 서비스 인터페이스
  - IBoardsServiceDeleteBoard에 userId 추가

feat-#18

+++ src/apis/boards/boards.controller.ts
@@ -72,13 +72,15 @@ export class BoardsController {
 
 	/**
 	 * DELETE '/boards/:id' 라우트 핸들러
+	 * @param req HTTP 요청 객체 - req.user: id, exp, role, nickname
 	 * @param id 게시글 id
 	 */
 	@Delete('/:id')
 	@UseGuards(restAuthGuard('access'))
 	deleteBoard(
-		@Param('id', ParseIntPipe) id: number, //
+		@Req() req: Request & IAuthUser, //
+		@Param('id', ParseIntPipe) id: number,
 	): Promise<void> {
-		return this.boardsService.deleteBoard({ id });
+		return this.boardsService.deleteBoard({ userId: req.user.id, id });
 	}
 }
